requirements.txt
allure-pytest
loguru
xlrd